\name{obs}
\alias{obs}
\title{Daily temperature and precipitation from Oslo}
\description{Daily temperature and precipitation from Oslo. Stored as a 'daily.station.record' (clim.pact).}
\usage{data(obs)}
\format{The data is a 'daily.station.record' object (see \code{\link[clim.pact]{station.obj}}).}
\examples{
data(obs)
}
\source{}
\references{
}
\keyword{datasets}
